requirements.txt
discord.py
pydotenv
peewee